refactor(circuit): extract HamiltonianAssemblyMixin from CircuitRoutines [Clean Code: SRP]
Tier 5b-A. Move the 37-method "build operators -> assemble Hamiltonian
-> compute eigensystem" cluster from circuit_internals/routines.py
into a new sibling mixin circuit_internals/hamiltonian_assembly.py
(1346 lines).  CircuitRoutines now subclasses both SubsystemTreeMixin
(Tier 5a) and HamiltonianAssemblyMixin.

The new mixin owns the full path from ``self.hamiltonian_symbolic``
to numerical output:

* per-variable operator-method construction
  (``_set_operators``, ``_generate_operator_methods``, the six
  ``_build_<extended|periodic>_operators_<basis>`` helpers, the
  ``_make_purely_harmonic_operator_method`` factory);
* Josephson cosine/sine evaluation
  (``_evaluate_matrix_cosine_terms``, ``_extract_junction_phase``,
  ``_term_has_cos_factor``/``_sin_factor``, ``_assemble_cos_term``/
  ``_sin_term``, ``_build_junction_phase_operator_list``);
* Hamiltonian assembly
  (``_hamiltonian_for_harmonic_extended_vars``,
  ``_hamiltonian_for_purely_harmonic``, ``_evaluate_hamiltonian``,
  ``hamiltonian``);
* eigenvalue / eigensystem (``_evals_calc``, ``_esys_calc``,
  ``_eigenvals_for_purely_harmonic``);
* symbolic-side preprocessing that produces the inputs above
  (``_diagonalize_purely_harmonic_hamiltonian``,
  ``_transform_hamiltonian``, ``_set_vars``, ``_set_vars_no_hd``);
* operator-method utilities (``identity_wrap_for_hd``,
  ``get_operator_by_name``, ``is_subsystem``, ``exp_i_operator``,
  ``offset_free_charge_values``, ``_set_harmonic_basis_osc_params``);
* the bridge helper ``_operator_from_sym_expr_wrapper`` consumed by
  ``SubsystemTreeMixin._update_interactions``.

The module-level ``_PERIODIC_OP_FUNCS`` lookup constant moved with
the methods that read it.

Cross-mixin attributes / callables consumed by the moved methods are
declared under ``TYPE_CHECKING`` on the new mixin so mypy resolves
``self.<x>`` correctly.  Two practical tweaks vs. byte-equivalent
relocation:

* ``CircuitRoutines._identity`` (the unbound method used as the
  ``"I_operator"`` factory in ``_generate_operator_methods``) is
  rewritten to ``type(self)._identity`` so the mixin doesn't need
  to import its own subclass at runtime.  Behaviour-equivalent for
  every concrete subclass (``Circuit`` / ``Subsystem``) since
  neither overrides ``_identity``.
* The borderline methods discussed in the planning doc are placed
  per the rule "consumer owns the data": ``_set_vars`` and the two
  Hamiltonian transforms go to the assembly mixin (operator
  factories consume them); ``cutoffs_dict`` /
  ``discretized_grids_dict_for_vars`` stay in the residual
  ``CircuitRoutines`` since they're cross-cutting Hilbert-space
  basics.

routines.py: 2217 -> 1041 lines (-53%).  Combined with Tier 5a, the
original 2447-line CircuitRoutines is now ~42% smaller.

Verified against the characterization-test goldens introduced in the
prior commit (4 ``hamiltonian()`` / ``eigenvals()`` fixtures + 5
lifecycle-dispatch tests, all passing at ``rtol=1e-10``).  Full
pytest: 322 passed, 14 skipped (the +13 vs. main are the new
characterization tests).  mypy clean (102 source files, +1 for the
new module).
